refactor: extract useNewTaskInput hook
